+++ RedPandaIDE/editor.cpp
@@ -5393,7 +5393,6 @@ void Editor::applySettings()
     setCaretColor(pSettings->editor().caretColor());
 
     codeFolding().indentGuides = pSettings->editor().showIndentLines();
-    codeFolding().indentGuidesColor = pSettings->editor().indentLineColor();
     codeFolding().fillIndents = pSettings->editor().fillIndents();
 
     QFont f=QFont();